add migrations

+++ main/models.py
@@ -35,7 +35,7 @@ class User(AbstractBaseUser):
     twitch_name = models.CharField(max_length=25, unique=True)
 
     banned = models.BooleanField(default=False)
-    type = models.IntegerField(choices=UserType.choices, default=UserType.USER)
+    type = models.PositiveSmallIntegerField(choices=UserType.choices, default=UserType.USER)
 
     blocks_placed = models.PositiveIntegerField(default=0)
     last_placement = models.PositiveBigIntegerField(default=0)
